refactor(config): rename toml model.rs to common.rs, glob re-export
Replace the fragile hand-maintained `model` re-export list in mod.rs with a
`pub(crate) use common::*` glob so moving items between the stage modules no
longer needs re-export edits.
